add some first MLM code
