+++ roles/server/molecule/mlm/converge.yml
@@ -1,0 +1,26 @@
+---
+- name: Converge machines
+  hosts: all
+  become: true
+  pre_tasks:
+    - name: Fix FQDN
+      ansible.builtin.set_fact:
+        ansible_fqdn: mlm.podman.loc
+
+  roles:
+    - role: stdevel.uyuni.server
+      uyuni_scc_reg_code:
+        - "<insert SLES code here>"
+        - "<insert MLM code here>"
+      uyuni_scc_mail: "<insert SCC mail here>"
+      uyuni_enable_monitoring: true
+      uyuni_allow_unsupported: true
+      # uyuni_channels:
+      #   - name: almalinux9
+      #     arch: x86_64
+      #   - name: almalinux9-appstream
+      #     arch: x86_64
+      #   - name: almalinux9-uyuni-client
+      #     arch: x86_64
+      # # uyuni_release: '2023.03'
+      # uyuni_firewall_config: true
