+++ roles/server/molecule/mlm/destroy.yml
@@ -1,0 +1,24 @@
+- name: Destroy molecule containers
+  hosts: molecule
+  gather_facts: false
+  tasks:
+    - name: Stop and remove container
+      delegate_to: localhost
+      containers.podman.podman_container:
+        name: "{{ inventory_hostname }}"
+        state: absent
+        rm: true
+    - name: Remove potentially stopped container
+      delegate_to: localhost
+      ansible.builtin.command:
+        cmd: podman container rm --ignore {{ inventory_hostname }}
+      changed_when: false
+
+- name: Remove dynamic molecule inventory
+  hosts: localhost
+  gather_facts: false
+  tasks:
+    - name: Remove dynamic inventory file
+      ansible.builtin.file:
+        path: "{{ molecule_ephemeral_directory }}/inventory/molecule_inventory.yml"
+        state: absent
