+++ roles/server/molecule/mlm/tasks/create-fail.yml
@@ -1,0 +1,14 @@
+- name: Retrieve container log
+  ansible.builtin.command:
+    cmd: >-
+      {% raw %}
+      podman logs
+      {% endraw %}
+      {{ item.stdout_lines[0] }}
+    # podman inspect --format='{{.HostConfig.LogConfig.Path}}'
+  changed_when: false
+  register: logfile_cmd
+
+- name: Display container log
+  ansible.builtin.fail:
+    msg: "{{ logfile_cmd.stderr }}"
