--- roles/server/molecule/suma/converge.yml
@@ -1,31 +1,0 @@
----
-- name: Converge machines
-  hosts: all
-  become: true
-  pre_tasks:
-    - name: Set hostname
-      block:
-        - name: Set (short) hostname
-          ansible.builtin.hostname:
-            name: suma
-
-        - name: Update /etc/hosts
-          ansible.builtin.lineinfile:
-            path: /etc/hosts
-            regexp: '^127.0.0.1'
-            line: '127.0.0.1 suma.vagrant.loc suma'
-
-  roles:
-    - role: stdevel.uyuni.server
-      uyuni_scc_reg_code: "<insert code here>"
-      uyuni_scc_mail: "<insert SCC mail here>"
-      uyuni_channels:
-        - name: almalinux9
-          arch: x86_64
-        - name: almalinux9-appstream
-          arch: x86_64
-        - name: almalinux9-uyuni-client
-          arch: x86_64
-      uyuni_enable_monitoring: true
-      uyuni_install_monitoring_formulas: true
-      uyuni_firewall_config: true
