--- roles/server/molecule/suma/molecule.yml
@@ -1,18 +1,0 @@
----
-dependency:
-  name: galaxy
-driver:
-  name: vagrant
-lint: |
-    yamllint .
-    ansible-lint
-    flake8
-platforms:
-  - name: suma4
-    box: sles15-sp5
-    cpus: 4
-    memory: 16384
-provisioner:
-  name: ansible
-verifier:
-  name: testinfra
