--- roles/server/molecule/suma/vars/main.yml
@@ -1,35 +1,0 @@
----
-# packages
-uyuni_core_packages:
-  - xfsprogs
-  - lvm2
-  - man
-uyuni_repos:
-  - "https://download.opensuse.org/repositories/systemsmanagement:/Uyuni:/Stable/images/repo/Uyuni-Server-POOL-x86_64-Media1/"
-  - "https://download.opensuse.org/repositories/systemsmanagement:/Uyuni:/Stable:/Patches/openSUSE_Leap_{{ ansible_distribution_version }}/systemsmanagement:Uyuni:Stable:Patches.repo"
-uyuni_packages:
-  - patterns-suma_server
-  - spacewalk-utils
-  - spacecmd
-
-# Uyuni configuration
-uyuni_org_name: "Demo"
-uyuni_org_login: "admin"
-uyuni_org_password: "admin"
-
-# firewall settings
-uyuni_firewall_config: false
-uyuni_firewall_default_zone: internal
-uyuni_firewall_services:
-  - suse-manager-server
-
-# content settings
-uyuni_channels:
-  - name: centos7
-    arch: x86_64
-  - name: centos7-updates
-    arch: x86_64
-
-# monitoring settings
-uyuni_enable_monitoring: true
-
